renaming entity for improving name convention as nest suggestion
